requirements.txt
Flask
Flask_SQLAlchemy
pycryptodome
